Added common notification timeout function and used it across the test runs page notifications
