+++ galasa-ui/src/components/common/UseNotification.tsx
@@ -1,0 +1,21 @@
+import { WARNING_NOTIFICATION_VISIBLE_MILLISECS } from '@/utils/constants/common';
+import { useEffect, useState } from 'react';
+
+export const useNotification = (initialState: boolean) => {
+  const [showNotification, setShowNotification] = useState(initialState);
+
+  useEffect(() => {
+    // Only show the notification if the limit was exceeded
+    if (initialState) {
+      setShowNotification(true);
+      const timer = setTimeout(() => {
+        setShowNotification(false);
+      }, WARNING_NOTIFICATION_VISIBLE_MILLISECS);
+
+      // Cleanup function: This will clear the timer on unmount.
+      return () => clearTimeout(timer);
+    }
+  }, [initialState]);
+
+  return showNotification;
+};
